remote-tmux: linear placement chrome, readiness-driven initial sizing, debug verbs isolated
Placement previously re-ran the recursive chrome fold for every child at
every level, walking each subtree once per ancestor; a one-pass ChromeTree
now threads each node's chrome through place(), keeping the derivation
linear in pane count.

The single-pane initial-sizing retry (20x sleep loop re-armed by two
NotificationCenter observers) is replaced by direct surface events: a new
TerminalSurface.onRuntimeReady callback fires the moment the runtime surface
becomes live — the one event guaranteed to happen exactly once even for a
surface created already AT its final grid, which never applies a resize and
so can never trigger a report-based hook (the deadlock the old polling loop
was papering over, reproduced 1/5 vs 5/5 in an A/B against the identical
machine state). The applied-size report stays as the update path, including
the off-window flush for background workspaces. Both hooks clear when a
window mirror takes ownership.

The DEBUG-only test_exec/test_set_frame socket verbs move to a dedicated
debug-only file: they exist because the sandboxed XCUITest runner cannot
create /tmp dirs, spawn a tmux server, or resize windows without AX
gestures, while the unsandboxed app can — a process boundary @testable
import cannot cross.
